requirements.txt
Flask==2.3.0
google-generativeai==0.2.0
gunicorn==20.1.0
